Move Nonce to its own file
