+++ pages/key-manager.tsx
@@ -2,9 +2,7 @@ import ERC725 from '@erc725/erc725.js';
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import { useState } from 'react';
-import useWeb3 from '../hooks/useWeb3';
-
-import LSP6KeyManager from '../abis/LSP6KeyManager.json';
+import KeyManagerNonceChecker from '../components/KeyManagerNonceChecker';
 
 const KeyManager: NextPage = () => {
   const [encodedPermissions, setEncodedPermissions] = useState(
@@ -49,6 +47,7 @@ const KeyManager: NextPage = () => {
         <title>Key Manager - ERC725 Tools</title>
       </Head>
       <div className="container">
+        <h2 className="title is-2">Permissions</h2>
         <article className="message is-info">
           <div className="message-body">
             This tool will encode/decode permissions according to{' '}
@@ -144,125 +143,8 @@ const KeyManager: NextPage = () => {
         </div>
       </div>
 
-      <Nonce />
+      <KeyManagerNonceChecker />
     </>
-  );
-};
-
-const Nonce = () => {
-  const web3 = useWeb3();
-
-  const [keyManagerAddress, setKeyManagerAddress] = useState('');
-  const [callerAddress, setCallerAddress] = useState('');
-  const [channelId, setChannelId] = useState('');
-  const [nonce, setNonce] = useState('');
-
-  const [showNonce, setShowNonce] = useState(false);
-  const [error, showError] = useState(false);
-
-  const getNonceFromKeyManager = async () => {
-    if (!web3) return;
-
-    const keyManagerInstance = new web3.eth.Contract(
-      LSP6KeyManager.abi as any,
-      web3.utils.toChecksumAddress(keyManagerAddress),
-    );
-
-    const isKeyManagerV05 = await keyManagerInstance.methods
-      .supportsInterface('0x6f4df48b')
-      .call();
-
-    const isKeyManagerV06 = await keyManagerInstance.methods
-      .supportsInterface('0xc403d48f')
-      .call();
-
-    if (isKeyManagerV05 || isKeyManagerV06) {
-      const result: any = await keyManagerInstance.methods
-        .getNonce(callerAddress, channelId)
-        .call();
-
-      setNonce(result);
-      setShowNonce(true);
-      showError(false);
-    } else {
-      setShowNonce(false);
-      showError(true);
-    }
-  };
-
-  return (
-    <div className="container mt-5">
-      <h1 className="title is-2">Nonce</h1>
-      <p>
-        Retrieve the nonce of an <code>address</code> for a specific Key
-        Manager.
-      </p>
-      <div className="columns mt-2">
-        <div className="column">
-          <div className="field">
-            <p className="control has-icons-left">
-              <input
-                className="input"
-                type="text"
-                placeholder="Key Manager address"
-                onChange={(e) => {
-                  setKeyManagerAddress(e.target.value);
-                }}
-              />
-              <span className="icon is-small is-left">🔐</span>
-            </p>
-          </div>
-        </div>
-        <div className="column">
-          <div className="field">
-            <p className="control has-icons-left">
-              <input
-                className="input"
-                type="text"
-                placeholder="Caller address"
-                onChange={(e) => setCallerAddress(e.target.value)}
-              />
-              <span className="icon is-small is-left">➡️ </span>
-            </p>
-          </div>
-        </div>
-        <div className="column">
-          <div className="field">
-            <p className="control has-icons-left">
-              <input
-                className="input"
-                type="text"
-                placeholder="Channel ID"
-                onChange={(e) => setChannelId(e.target.value)}
-              />
-              <span className="icon is-small is-left">🚪</span>
-            </p>
-          </div>
-        </div>
-        <div className="column">
-          <button
-            className="button is-success"
-            onClick={async () => {
-              await getNonceFromKeyManager();
-            }}
-          >
-            Get Nonce
-          </button>
-        </div>
-      </div>
-      <div
-        className="notification is-danger is-light"
-        style={{ display: error ? 'block' : 'none' }}
-      >
-        address <code>{keyManagerAddress}</code> is not a KeyManager
-      </div>
-      <div
-        className="notification is-success is-light"
-        style={{ display: showNonce ? 'block' : 'none' }}
-      >
-        Nonce = <code>{nonce}</code>
-      </div>
-    </div>
   );
 };
